test step: poison the LLL causing runahead
this poisons only the LLL causing entry to runahead without completing it immediately. the runahead exit condition is getting a packet response from memory. since the LLL still blocks the ROB, the only executed instruction in runahead is the LLL that caused entry into it and the only poisoned register is the LLL's destination register.