Arch checkpoint restoration
Uses a predetermined rename map, scoreboard and free list which are copied into the CPU when runahead has finished squashing after an exit. This approach introduces some memory/pointer bookkeeping but seems to work alright. After these structures are copied, checkpointed values are restored in the register file using the predetermined rename map. The only misc. register to be restored is the flags register. This is kind of hacky as it's specifically targeted for the X86 ISA.