new dynamic exit
kind of a hybrid between what is now NLLB and minimum work. use NLLB, but limit it to a certain number of instructions (also controlled by the minimum work param). if the load is too far off, the CPU checks if there is a runahead chain. if so, the chain executes as many times as it can fit within the minimum work limit, otherwise it exits immediately